+++ src/database.h
@@ -169,7 +169,7 @@ class DBResult
 		}
 
 		std::string getString(const std::string& s) const;
-		const char* getStream(const std::string& s, unsigned long& size) const;
+		const char* getStream(const std::string& s, unsigned long long& size) const;
 
 		bool hasNext() const;
 		bool next();
